requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.3
django-browser-reload==1.13.0
django-chartjs==2.3.0
django-otp==1.5.0
django-rest-auth==0.9.5
django_chart_js==0.0.3
djangorestframework==3.15.1
et-xmlfile==1.1.0
fonttools==4.53.1
google-auth==2.30.0
html5lib==1.1
idna==3.7
oauthlib==3.2.2
openpyxl==3.1.5
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.15.0
python-decouple==3.8
python-social-auth==0.3.6
python3-openid==3.2.0
razorpay==1.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.2.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tinycss2==1.3.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
zopfli==0.2.3
